fix(widget): wrap useSearchParams in Suspense boundary for static prerender
Next.js 15 requires useSearchParams() to be inside a Suspense boundary
for static generation. Split widget/apply into a server page wrapper
and a client content component.
